Day 10: simplify the loop-finding loop
Previously, findNext would return null for "S", as it's not a pipe,
(causing us to break out of the loop in findLoop), and its coordinate
was still being added to findLoop's resulting list but with a wrong
direction, which was just repeating the last known one.

Now, coordinate of the pipe animal stands on is still added, but with
the actual direction that lets to go there from the last pre-animal
pipe.